Refactor: 참가자 정보 type 정리
