+++ components/agenda/agendaDetail/AgendaTab.tsx
@@ -1,15 +1,8 @@
 import { useState } from 'react';
-import { AgendaDataProps } from 'types/agenda/agendaDetail/agendaDataTypes';
-import { TeamDataProps } from 'types/agenda/team/teamDataTypes';
+import { AgendaTabProps } from 'types/agenda/agendaDetail/tabs/agendaInfoTypes';
 import { TABS, TabValues } from 'constants/agenda/agendaDetail/agendaTabs';
 import TabContent from 'components/agenda/agendaDetail/TabContent';
 import TabButtons from 'components/agenda/button/TabButtons';
-
-interface AgendaTabProps {
-  agendaData: AgendaDataProps;
-  isHost: boolean;
-  myTeam?: TeamDataProps | null;
-}
 
 export default function AgendaTab({
   agendaData,
