+++ hooks/agenda/agendaDetail/useAgendaData.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import { AgendaDataProps } from 'types/agenda/agendaDetail/agendaDataTypes';
+import { AgendaDataProps } from 'types/agenda/agendaDetail/agendaTypes';
 import { instanceInAgenda } from 'utils/axios';
 
 export const useAgendaData = (agendaKey: string) => {
