+++ types/agenda/agendaDetail/tabs/agendaInfoTypes.ts
@@ -1,9 +1,14 @@
-import { AgendaStatus } from 'constants/agenda/agenda';
+import { TeamDataProps } from 'types/agenda/team/teamDataTypes';
+import { AgendaDataProps } from '../agendaTypes';
 
-// 참가자 정보 요약
-export interface ParticipantSummaryProps {
-  agendaStatus: AgendaStatus;
+export interface AgendaInfoProps {
+  agendaData: AgendaDataProps;
   isHost: boolean;
   isParticipant: boolean;
-  isTeam: boolean;
+}
+
+export interface AgendaTabProps {
+  agendaData: AgendaDataProps;
+  isHost: boolean;
+  myTeam?: TeamDataProps | null;
 }
